requirements.txt
python-dotenv
requests
git+https://github.com/narmitech/banking-client-python.git
